--- tests/test_5_celdas_por_fila.py
@@ -1,5 +1,0 @@
-from src import bingo
-mi_carton=bingo.crear_carton()
-
-def test_cinco_celdas_por_fila():
-    assert  bingo.cinco_celdas_por_fila(mi_carton)
