--- tests/test_columnas_sin_tres_celdas.py
@@ -1,6 +1,0 @@
-from src import bingo
-
-mi_carton = bingo.crear_carton()
-
-def test_columnas_sin_tres_celdas():
-    assert bingo.columnas_sin_tres_celdas(mi_carton)
