--- tests/test_filas.py
@@ -1,4 +1,0 @@
-from src import bingo
-mi_carton = bingo.crear_carton()
-def test_filas_no_vacias():
-    assert bingo.filas_no_vacias( mi_carton )
